fix: download cid or ipfs path to fs (api changes)
License: MIT
