+++ package.json
@@ -89,13 +89,13 @@
     "ipfs-http-client": "55.0.0",
     "ipfsd-ctl": "10.0.6",
     "is-ipfs": "6.0.2",
-    "it-all": "^1.0.6",
-    "it-concat": "^2.0.0",
     "it-last": "^1.0.6",
+    "it-pipe": "^1.1.0",
     "multiaddr": "10.0.1",
     "multiaddr-to-uri": "8.0.0",
     "portfinder": "^1.0.28",
     "recursive-readdir": "^2.2.2",
+    "stream-to-it": "^0.2.4",
     "stream-to-pull-stream": "^1.7.3",
     "sudo-prompt": "^9.2.1",
     "untildify": "^4.0.0",
